src/cn/seu/edu/spring/chapter4/soundsystem/TrackCounter.java
package cn.seu.edu.spring.chapter4.soundsystem;

import org.aspectj.lang.annotation.Aspect;
import org.aspectj.lang.annotation.Before;
import org.aspectj.lang.annotation.Pointcut;

import java.util.HashMap;
import java.util.Map;

/**
 * Created by Administrator on 2017/9/17.
 */
@Aspect
public class TrackCounter {
    private Map<Integer, Integer> trackCounts = new HashMap<>();

    @Pointcut("execution(* cn.seu.edu.spring.chapter4.soundsystem.BlankDisc.play(int))" +
            "&& args(trackNumber)")
    public void play(int trackNumber){}

    @Before("play(trackNumber)")
    public void countTrack(int trackNumber) {
        int currentCount = getPlayCount(trackNumber);
        trackCounts.put(trackNumber, currentCount + 1);
    }

    public int getPlayCount(int trackNumber) {
        return trackCounts.containsKey(trackNumber) ? trackCounts.get(trackNumber) : 0;
    }
}
